backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
qdrant-client==1.7.0
pydantic==2.5.0
langchain==0.2.16
langchain-ollama==0.1.0
langchain-google-genai==1.0.7
langgraph>=0.2.0
google-generativeai>=0.5.2
# google-generativeai==0.8.3
